Внедрена проверка роли пользователей. (#90)
* Внедрена проверка роли пользователей.

* Исправлены тесты, сборка собирается.

* Update OrderServiceImpl.java

Испрвлено DTO на Dto

+++ src/main/java/com/homegarden/store/backend/service/CartServiceImpl.java
@@ -17,10 +17,13 @@ public class CartServiceImpl implements CartService {
     private final CartRepository cartRepository;
     private final UserService userService;
 
+    private final AccessCheckService accessCheckService;
+
     @Override
     public Cart create(Cart cart) {
         Long userId = cart.getUser().getUserId();
         User user = userService.getById(userId);
+        accessCheckService.checkAccess(user);
         if(cartRepository.existsCartByUser(user)) {
             throw new CartAlreadyExistsException("Cart already exists for this user");
         }
@@ -30,7 +33,10 @@ public class CartServiceImpl implements CartService {
 
     @Override
     public Cart getById(Long id) {
-        return cartRepository.findById(id).orElseThrow(() -> new CartNotFoundException("Cart not found"));
+        Cart cart = cartRepository.findById(id)
+                .orElseThrow(() -> new CartNotFoundException("Cart not found"));
+        accessCheckService.checkAccess(cart);
+        return cart;
     }
 
     @Override
@@ -40,7 +46,8 @@ public class CartServiceImpl implements CartService {
 
     @Override
     public void delete(Long id) {
-        getById(id);
+        Cart cart = getById(id);
+        accessCheckService.checkAccess(cart);
         cartRepository.deleteById(id);
     }
